refactor: move env validation to dedicated utility module - Fixes #238
