+++ back-end/server.js
@@ -11,17 +11,8 @@ const mongoose = require('mongoose');
 const PORT = process.env.PORT || 5000;
 
 // ==================== ENVIRONMENT VALIDATION ====================
-const requiredEnvVars = ['PORT', 'MONGODB_URI'];
-const missingEnvVars = requiredEnvVars.filter(
-  (envVar) => !process.env[envVar]
-);
-
-if (missingEnvVars.length > 0) {
-  console.error(
-    `Missing required environment variables: ${missingEnvVars.join(', ')}`
-  );
-  process.exit(1);
-}
+const validateEnv = require("./src/config/validateEnv");
+validateEnv();
 
 // ==================== UNHANDLED REJECTIONS & EXCEPTIONS ====================
 process.on('unhandledRejection', (reason, promise) => {
